Move client and exceptions and prepare for new api clients
